LayoutTests/platform/chromium/http/tests/inspector/console-websocket-error-expected.txt

CONSOLE MESSAGE: line 32: testDNSLookup: Test started.
CONSOLE MESSAGE: WebSocket network error: The operation couldn’t be completed. (kCFErrorDomainCFNetwork error 2.)
CONSOLE MESSAGE: line 43: testDNSLookup: onclose is called.
CONSOLE MESSAGE: line 50: testSSLCertificate: Test started.
CONSOLE MESSAGE: WebSocket network error: OSStatus Error -9812: valid cert chain, untrusted root 
CONSOLE MESSAGE: line 64: testSSLCertificate: onclose is called.
WebSocket's network errors should be logged to console.

console-websocket-error.html:32testDNSLookup: Test started.
ws://nonexistent.domain.invalid/WebSocket network error: The operation couldn’t be completed. (kCFErrorDomainCFNetwork error 2.)
console-websocket-error.html:43testDNSLookup: onclose is called.
console-websocket-error.html:50testSSLCertificate: Test started.
wss://127.0.0.1:8443/WebSocket network error: OSStatus Error -9812: valid cert chain, untrusted root 
console-websocket-error.html:64testSSLCertificate: onclose is called.
